+++ src/Components/Meals/Meals.js
@@ -2,45 +2,7 @@ import React, { useState } from "react";
 import Meal from "./Meal";
 
 import classes from "./Meals.module.css";
-
-const meals = [
-  {
-    id: 1,
-    mealName: "Sushi",
-    mealExp: "Finest fish and veggies",
-    mealPrice: "22.99",
-    total: +0,
-    totalAmount: +0,
-    isAdded: false,
-  },
-  {
-    id: 2,
-    mealName: "Schnitzel",
-    mealExp: "A german speciality",
-    mealPrice: "16.50",
-    total: +0,
-    totalAmount: +0,
-    isAdded: false,
-  },
-  {
-    id: 3,
-    mealName: "Barbecue Burger",
-    mealExp: "American, raw, meaty",
-    mealPrice: "12.99",
-    total: +0,
-    totalAmount: +0,
-    isAdded: false,
-  },
-  {
-    id: 4,
-    mealName: "Green Bowl",
-    mealExp: "Healthy...and gree...",
-    mealPrice: "18.99",
-    total: +0,
-    totalAmount: +0,
-    isAdded: false,
-  },
-];
+import meals from "../../constants/Meals";
 
 const Meals = (props) => {
   const [newMeal, setNewMeal] = useState(meals);
